+++ src/axom/core/utilities/FileUtilities.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -29,6 +29,9 @@
   #define Stat stat
   #define Unlink unlink
 #endif
+
+// Note: The hard-wired path separator character in this file
+// should be set to the backslash when on Windows.
 
 namespace axom
 {
@@ -104,6 +107,53 @@ int makeDirsForPath(const std::string& path)
   } while(pos != std::string::npos);
 
   return err;
+}
+
+//-----------------------------------------------------------------------------
+std::string prefixRelativePath(const std::string& path, const std::string& prefix)
+{
+  if(path.empty())
+  {
+    throw std::invalid_argument("path must not be empty");
+  };
+  if(path[0] == '/' || prefix.empty())
+  {
+    return path;
+  }
+  return utilities::filesystem::joinPath(prefix, path);
+}
+
+//-----------------------------------------------------------------------------
+std::string getParentPath(const std::string& path)
+{
+  if(path.empty())
+  {
+    throw std::invalid_argument("path must not be empty");
+  };
+
+  char separator = '/';
+
+  std::string parent;
+
+  if(path.size() == 1 && path[0] == separator)
+  {
+    // path is root, so parent is blank.
+  }
+  else
+  {
+    std::size_t found = path.rfind(separator);
+
+    if(found != std::string::npos)
+    {
+      if(found == 0)
+      {
+        ++found;
+      }
+      parent = path.substr(0, found);
+    }
+  }
+
+  return parent;
 }
 
 //-----------------------------------------------------------------------------