+++ src/axom/core/utilities/FileUtilities.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -65,6 +65,30 @@ std::string joinPath(const std::string& fileDir,
  * doesn't exist, it is created.
  */
 int makeDirsForPath(const std::string& path);
+
+/*!
+ * \brief Add a prefix to a path if it is relative.
+ *
+ * \param [in] path non-empty string representing an absolute or relative path
+ * \param [in] prefix string representing a directory path
+ *
+ * \return \c prefix + \c path (with delimiter) if the path is
+ * relative or \c prefix is empty, or just \c path otherwise.
+ */
+std::string prefixRelativePath(const std::string& path,
+                               const std::string& prefix);
+
+/*!
+ * \brief Get parent path name from a filesystem path.
+ *
+ * \param [in] path a non-empty absolute or relative filesystem path
+ *
+ * \return a directory path formed by removing the last part of the
+ *    input path
+ *
+ * If path is the system root, its parent is the empty string.
+ */
+std::string getParentPath(const std::string& path);
 
 /*!
  * \brief Get directory name from a path that contains a file name