+++ src/axom/primal/geometry/KnotVector.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -157,8 +157,8 @@ public:
    */
   T& operator[](axom::IndexType i)
   {
-    return m_knots[i];
     SLIC_ASSERT(isValid());
+    return m_knots[i];
   }
 
   /// \brief Return the degree of the knot vector