+++ src/axom/quest/Shaper.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -16,14 +16,23 @@
 #ifndef AXOM_USE_KLEE
   #error Shaping functionality requires Axom to be configured with the Klee component
 #endif
-#ifndef AXOM_USE_MFEM
-  #error Shaping functionality requires Axom to be configured with MFEM and the AXOM_ENABLE_MFEM_SIDRE_DATACOLLECTION option
+
+#if !defined(AXOM_USE_MFEM) && !defined(AXOM_USE_CONDUIT)
+  #error Shaping functionality requires Axom to be configured with Conduit or MFEM and the AXOM_ENABLE_MFEM_SIDRE_DATACOLLECTION option
 #endif
 
 #include "axom/sidre.hpp"
 #include "axom/klee.hpp"
 #include "axom/mint.hpp"
 #include "axom/quest/DiscreteShape.hpp"
+#include "axom/core/execution/runtime_policy.hpp"
+
+#if defined(AXOM_USE_MFEM)
+  #include "mfem.hpp"
+#endif
+#if defined(AXOM_USE_CONDUIT)
+  #include "conduit_node.hpp"
+#endif
 
 #include "axom/quest/interface/internal/mpicomm_wrapper.hpp"
 
@@ -33,13 +42,51 @@ namespace quest
 {
 /**
  * Abstract base class for shaping material volume fractions
+ *
+ * Shaper requires Axom to be configured with Conduit or MFEM
+ * or both.
  */
 class Shaper
 {
 public:
-  Shaper(const klee::ShapeSet& shapeSet, sidre::MFEMSidreDataCollection* dc);
+  using RuntimePolicy = axom::runtime_policy::Policy;
 
-  virtual ~Shaper() = default;
+#if defined(AXOM_USE_MFEM)
+  /*!
+    @brief Construct Shaper to operate on an MFEM mesh.
+  */
+  Shaper(RuntimePolicy execPolicy,
+         int allocatorId,
+         const klee::ShapeSet& shapeSet,
+         sidre::MFEMSidreDataCollection* dc);
+#endif
+
+  /*!
+    @brief Construct Shaper to operate on a blueprint-formatted mesh
+    stored in a sidre Group.
+  */
+  Shaper(RuntimePolicy execPolicy,
+         int allocatorId,
+         const klee::ShapeSet& shapeSet,
+         sidre::Group* bpMesh,
+         const std::string& topo = "");
+
+  /*!
+    @brief Construct Shaper to operate on a blueprint-formatted mesh
+    stored in a conduit Node.
+
+    Because \c conduit::Node doesn't support application-specified
+    allocator id for (only) arrays, the incoming \c bpNode must have
+    all arrays pre-allocated in a space accessible by the runtime
+    policy.  Any needed-but-missing space would lead to an exception.
+  */
+  Shaper(RuntimePolicy execPolicy,
+         int allocatorId,
+         const klee::ShapeSet& shapeSet,
+         conduit::Node& bpNode,
+         const std::string& topo = "");
+
+  virtual ~Shaper();
 
 public:
   // Some default values.
@@ -50,6 +97,9 @@ public:
 
   /// Refinement type.
   using RefinementType = DiscreteShape::RefinementType;
+
+  //! @brief Verify the input mesh is okay for this class to work with.
+  bool verifyInputMesh(std::string& whyBad) const;
 
   //@{
   //!  @name Functions to get and set shaping parameters
@@ -62,10 +112,22 @@ public:
 
   //@}
 
+  /*!
+    @brief Set path of shape input file.
+
+    The path is used to resolve relative paths that may have been
+    specified in the file.
+  */
+  void setFilePath(const std::string& filePath);
+
+  mint::Mesh* getSurfaceMesh() const { return m_surfaceMesh.get(); }
+
   bool isVerbose() const { return m_verboseOutput; }
 
+#ifdef AXOM_USE_MFEM
   sidre::MFEMSidreDataCollection* getDC() { return m_dc; }
-  mint::Mesh* getSurfaceMesh() const { return m_surfaceMesh.get(); }
+  const sidre::MFEMSidreDataCollection* getDC() const { return m_dc; }
+#endif
 
   /*!
    * \brief Predicate to determine if the specified format is valid
@@ -152,10 +214,34 @@ protected:
   int getRank() const;
 
 protected:
+  RuntimePolicy m_execPolicy;
+  int m_allocatorId;
+
+  // For any mesh represented in Conduit or sidre
   sidre::DataStore m_dataStore;
 
   const klee::ShapeSet& m_shapeSet;
-  sidre::MFEMSidreDataCollection* m_dc;
+
+  //! \brief Prefix path for shape file names with relative path.
+  std::string m_prefixPath;
+
+#if defined(AXOM_USE_MFEM)
+  // For mesh represented as MFEMSidreDataCollection
+  sidre::MFEMSidreDataCollection* m_dc {nullptr};
+#endif
+
+#if defined(AXOM_USE_CONDUIT)
+  //! @brief Version of the mesh for computations.
+  axom::sidre::Group* m_bpGrp;
+  const std::string m_bpTopo;
+  //! @brief Mesh in an external Node, when provided as a Node.
+  conduit::Node* m_bpNodeExt;
+  //! @brief Initial copy of mesh in an internal Node storage.
+  conduit::Node m_bpNodeInt;
+#endif
+
+  //! @brief Number of cells in computational mesh (m_dc or m_bpGrp).
+  axom::IndexType m_cellCount;
 
   std::shared_ptr<mint::Mesh> m_surfaceMesh;