+++ src/axom/sidre/core/View.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
 // other Axom Project Developers. See the top-level LICENSE file for details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
@@ -826,6 +826,9 @@ public:
     //If debug, should trigger assert.  If release, issue warning.
     return getData();
   }
+
+  /// \overload
+  Node::ConstValue getArray() const { return getData(); }
 
   /*!
    * \brief Returns a pointer to the string contained in the view.